manifest: refine component manifest handling
