+++ pkg/repository/v1manifest/repo.go
@@ -18,11 +18,16 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"io"
 	"io/ioutil"
+	"os"
+	"path/filepath"
+	"strings"
 	"time"
 
 	cjson "github.com/gibson042/canonicaljson-go"
 	"github.com/pingcap-incubator/tiup/pkg/repository/crypto"
+	"github.com/pingcap/errors"
 )
 
 // Init creates and initializes an empty repository
@@ -78,6 +83,8 @@ func Init(dst string, initTime time.Time, priv string) error {
 
 // AddComponent adds a new component to an existing repository
 func AddComponent(id, name, desc, owner, repo string, isDefault bool, pub, priv string) error {
+	id = strings.ToLower(id)
+
 	// read key files
 	privBytes, err := ioutil.ReadFile(priv)
 	if err != nil {
@@ -95,19 +102,24 @@ func AddComponent(id, name, desc, owner, repo string, isDefault bool, pub, priv 
 	}
 
 	// check id conflicts
+	if _, found := ManifestsConfig[id]; found {
+		// reserved keywords
+		return fmt.Errorf("component id '%s' is not allowed, please use another one", id)
+	}
 	if _, found := manifests[ManifestTypeIndex].(*Index).Components[id]; found {
-		return fmt.Errorf("component id '%s' already exist, please set another one", id)
+		return fmt.Errorf("component id '%s' already exist, please use another one", id)
 	}
 
 	// create new component manifest
 	currTime := time.Now().UTC()
-	comp := NewComponent(name, desc, currTime)
+	comp := NewComponent(id, name, desc, currTime)
+	manifests[id] = comp
 
 	// update repository
 	compInfo := ComponentItem{
 		Owner:     owner,
-		URL:       fmt.Sprintf("/%s", comp.Filename()),
-		Threshold: 1, // TODO: make this configurable
+		URL:       comp.Filename(),
+		Threshold: 1,
 	}
 	index := manifests[ManifestTypeIndex].(*Index)
 	index.Components[id] = compInfo
@@ -121,11 +133,12 @@ func AddComponent(id, name, desc, owner, repo string, isDefault bool, pub, priv 
 	snapshot.Expires = currTime.Add(ManifestsConfig[ManifestTypeSnapshot].Expire).Format(time.RFC3339)
 
 	// update timestamp
-	timestamp, err := manifests[ManifestTypeTimestamp].(*Timestamp).SetSnapshot(snapshot)
+	timestamp, err := NewTimestamp(currTime).SetSnapshot(snapshot)
 	if err != nil {
 		return err
 	}
-	timestamp.Expires = currTime.Add(ManifestsConfig[ManifestTypeTimestamp].Expire).Format(time.RFC3339)
+	timestamp.Version = manifests[ManifestTypeTimestamp].(*Timestamp).Version + 1
+	manifests[ManifestTypeTimestamp] = timestamp
 
 	return BatchSaveManifests(repo, manifests, privKey)
 }
@@ -177,13 +190,13 @@ func NewTimestamp(initTime time.Time) *Timestamp {
 			Ty:          ManifestTypeTimestamp,
 			SpecVersion: CurrentSpecVersion,
 			Expires:     initTime.Add(ManifestsConfig[ManifestTypeTimestamp].Expire).Format(time.RFC3339),
-			Version:     0, // not versioned
+			Version:     1,
 		},
 	}
 }
 
 // NewComponent creates a Component object
-func NewComponent(name, desc string, initTime time.Time) *Component {
+func NewComponent(id, name, desc string, initTime time.Time) *Component {
 	return &Component{
 		SignedBase: SignedBase{
 			Ty:          ManifestTypeComponent,
@@ -191,6 +204,7 @@ func NewComponent(name, desc string, initTime time.Time) *Component {
 			Expires:     initTime.Add(ManifestsConfig[ManifestTypeComponent].Expire).Format(time.RFC3339),
 			Version:     1,
 		},
+		ID:          id,
 		Name:        name,
 		Description: desc,
 		Platforms:   make(map[string]map[string]VersionItem),
@@ -264,7 +278,7 @@ func (manifest *Root) SetRole(m ValidManifest) {
 	}
 
 	manifest.Roles[m.Base().Ty] = &Role{
-		URL:       fmt.Sprintf("/%s", m.Filename()),
+		URL:       m.Filename(),
 		Threshold: ManifestsConfig[m.Base().Ty].Threshold,
 		Keys:      make(map[string]*KeyInfo),
 	}
@@ -293,4 +307,77 @@ func FreshKeyInfo() (*KeyInfo, string, crypto.PrivKey, error) {
 	hash := sha256.Sum256(serInfo)
 
 	return &info, fmt.Sprintf("%x", hash), priv, nil
+}
+
+// ReadManifestDir reads manifests from a dir
+func ReadManifestDir(dir string) (map[string]ValidManifest, error) {
+	manifests := make(map[string]ValidManifest)
+	for ty, val := range ManifestsConfig {
+		if val.Filename == "" {
+			continue
+		}
+		reader, err := os.Open(filepath.Join(dir, val.Filename))
+		if err != nil {
+			return nil, err
+		}
+		defer reader.Close()
+		var role ValidManifest
+		m, err := ReadManifest(reader, role, crypto.NewKeyStore())
+		if err != nil {
+			return nil, err
+		}
+		manifests[ty] = m.Signed
+	}
+	return manifests, nil
+}
+
+// SignAndWrite creates a manifest and writes it to out.
+func SignAndWrite(out io.Writer, role ValidManifest, privKey crypto.PrivKey) error {
+	payload, err := cjson.Marshal(role)
+	if err != nil {
+		return err
+	}
+
+	sign, err := privKey.Signature(payload)
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	keyInfo, err := NewKeyInfo(privKey)
+	if err != nil {
+		return err
+	}
+	keyID, err := keyInfo.ID()
+	if err != nil {
+		return err
+	}
+
+	manifest := Manifest{
+		Signatures: []signature{{
+			KeyID: keyID,
+			Sig:   string(sign),
+		}},
+		Signed: role,
+	}
+
+	encoder := json.NewEncoder(out)
+	encoder.SetIndent("", "\t")
+	return encoder.Encode(manifest)
+}
+
+// BatchSaveManifests write a series of manifests to disk
+func BatchSaveManifests(dst string, manifestList map[string]ValidManifest, privKey crypto.PrivKey) error {
+	for _, m := range manifestList {
+		writer, err := os.OpenFile(filepath.Join(dst, m.Filename()), os.O_WRONLY|os.O_TRUNC|os.O_CREATE, 0644)
+		if err != nil {
+			return err
+		}
+		defer writer.Close()
+		// TODO: support multiples keys
+
+		if err = SignAndWrite(writer, m, privKey); err != nil {
+			return err
+		}
+	}
+	return nil
 }
